+++ src/AutoRest.CSharp/Mgmt/Decorator/CodeModelTransformer.cs
@@ -15,6 +15,11 @@ namespace AutoRest.CSharp.Mgmt.Decorator
             SchemaRenamer.UpdateAcronyms();
             UrlToUri.UpdateSuffix();
             BodyParameterNormalizer.UpdatePatchOperations();
+            FrameworkTypeUpdater.ValidateAndUpdate();
+            SealedChoicesUpdater.UpdateSealChoiceTypes();
+            CommonSingleWordModels.Update();
+            RenameTimeToOn.Update();
+            RearrangeParameterOrder.Update();
 
             CodeModelValidator.Validate();
         }
